# Reduce string memory allocation when template save